backend/requirements.txt
Django==2.0.6
django-mysql==2.4.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.3
mysqlclient==1.3.13
python-decouple==3.1
